refractor: VCursor/VSchema file splitting
